--- apps/financial-aid/web-veita/src/components/AdminLayout/index.ts
@@ -1,1 +1,0 @@
-export { default as AdminLayout } from './AdminLayout'
