--- apps/financial-aid/web-veita/src/components/AidAmountModal/index.ts
@@ -1,1 +1,0 @@
-export { default as AidAmountModal } from './AidAmountModal'
