--- apps/financial-aid/web-veita/src/components/ApplicationsTable/index.ts
@@ -1,1 +1,0 @@
-export { default as ApplicationsTable } from './ApplicationsTable'
