--- apps/financial-aid/web-veita/src/components/Generator/index.ts
@@ -1,2 +1,0 @@
-export { default as GeneratedProfile } from './GeneratedProfile'
-export { default as GenerateName } from './GenerateName'
