--- apps/financial-aid/web-veita/src/components/History/index.ts
@@ -1,1 +1,0 @@
-export { default as History } from './History'
