--- apps/financial-aid/web-veita/src/components/InputModal/index.ts
@@ -1,1 +1,0 @@
-export { default as InputModal } from './InputModal'
