+++ apps/financial-aid/web-veita/src/components/Nav/Nav.tsx
@@ -8,9 +8,10 @@ import {
   SkeletonLoader,
 } from '@island.is/island-ui/core'
 import { useRouter } from 'next/router'
-import Link from 'next/link'
 
 import {
+  AdminSideNavItems,
+  EmployeeSideNavItems,
   LoadingContainer,
   LogoMunicipality,
 } from '@island.is/financial-aid-web/veita/src/components'
@@ -20,13 +21,6 @@ import cn from 'classnames'
 import { ApplicationFiltersContext } from '@island.is/financial-aid-web/veita/src/components/ApplicationFiltersProvider/ApplicationFiltersProvider'
 
 import { useLogOut } from '@island.is/financial-aid-web/veita/src/utils/useLogOut'
-import {
-  ApplicationFiltersEnum,
-  Routes,
-  StaffRole,
-} from '@island.is/financial-aid/shared/lib'
-
-import { navigationItems } from '@island.is/financial-aid-web/veita/src/utils/navigation'
 
 import { AdminContext } from '@island.is/financial-aid-web/veita/src/components/AdminProvider/AdminProvider'
 
@@ -72,59 +66,16 @@ const Nav = ({ showInMobile }: Props) => {
           isLoading={loading}
           loader={<SkeletonLoader repeat={3} space={2} />}
         >
-          {admin?.staff?.roles.includes(StaffRole.EMPLOYEE) &&
-            navigationItems.map((item, index) => {
-              return (
-                <div key={'NavigationLinks-' + index}>
-                  {item.group && <p className={styles.group}>{item.group}</p>}
-                  <Link href={item.link}>
-                    <a
-                      aria-label={item.label}
-                      className={cn({
-                        [`${styles.link}`]: true,
-                        [`${styles.activeLink}`]: router.pathname === item.link,
-                        [`${styles.linkHoverEffect}`]:
-                          router.pathname !== item.link,
-                      })}
-                    >
-                      <Box display="flex" justifyContent="spaceBetween">
-                        <Text fontWeight="semiBold">{item.label}</Text>
-                        <Text fontWeight="semiBold" color="dark300">
-                          {item.applicationState
-                            .map((state: ApplicationFiltersEnum) => {
-                              if (applicationFilters) {
-                                return applicationFilters[state]
-                              }
-                            })
-                            .reduce((a?: number, b?: number) => {
-                              return (a || 0) + (b || 0)
-                            })}
-                        </Text>
-                      </Box>
-                    </a>
-                  </Link>
-                </div>
-              )
-            })}
+          <EmployeeSideNavItems
+            roles={admin?.staff?.roles}
+            applicationFilters={applicationFilters}
+          />
         </LoadingContainer>
       </div>
 
       <Box display="block" marginBottom={2} marginTop={4}>
         <Box marginBottom={2}>
-          {admin?.staff?.roles.includes(StaffRole.ADMIN) && (
-            <button
-              className={`${styles.sideNavBarButton} navBarButtonHover`}
-              onClick={() => router.push(Routes.users)}
-            >
-              <Icon
-                icon="people"
-                type="outline"
-                color="blue400"
-                className={styles.sideNavBarButtonIcon}
-              />
-              <Text> Notendur</Text>
-            </button>
-          )}
+          <AdminSideNavItems roles={admin?.staff?.roles} />
           <button
             className={`${styles.sideNavBarButton} navBarButtonHover`}
             onClick={() => logOut()}
