--- apps/financial-aid/web-veita/src/components/Nav/index.ts
@@ -1,1 +1,0 @@
-export { default as Nav } from './Nav'
