--- apps/financial-aid/web-veita/src/components/Profile/index.ts
@@ -1,1 +1,0 @@
-export { default as Profile } from './Profile'
