--- apps/financial-aid/web-veita/src/components/StateModal/index.ts
@@ -1,1 +1,0 @@
-export { default as StateModal } from './StateModal'
