--- apps/financial-aid/web-veita/src/components/TableBody/index.ts
@@ -1,1 +1,0 @@
-export { default as TableBody } from './TableBody'
