--- apps/financial-aid/web-veita/src/components/TableHeaders/index.ts
@@ -1,1 +1,0 @@
-export { default as TableHeaders } from './TableHeaders'
