+++ apps/financial-aid/web-veita/src/components/UsersTableBody/UsersTableBody.tsx
@@ -20,60 +20,58 @@ const UsersTableBody = ({ user, index }: PageProps) => {
   const { admin } = useContext(AdminContext)
   const isLoggedInUser = admin?.nationalId === user.nationalId
   return (
-    <>
-      <tr
-        className={`${styles.link} contentUp`}
-        style={{ animationDelay: 55 + 3.5 * index + 'ms' }}
+    <tr
+      className={`${styles.link} contentUp`}
+      style={{ animationDelay: 55 + 3.5 * index + 'ms' }}
+    >
+      <td
+        className={cn({
+          [`${styles.tablePadding} ${styles.firstChildPadding}`]: true,
+        })}
       >
-        <td
-          className={cn({
-            [`${styles.tablePadding} ${styles.firstChildPadding}`]: true,
-          })}
-        >
-          <Box display="flex">
-            <Text variant="h5">
-              {user.name} {isLoggedInUser ? '(Þú)' : ''}
-            </Text>
-          </Box>
-        </td>
+        <Box display="flex">
+          <Text variant="h5">
+            {user.name} {isLoggedInUser ? '(Þú)' : ''}
+          </Text>
+        </Box>
+      </td>
 
+      <td
+        className={cn({
+          [`${styles.tablePadding} `]: true,
+        })}
+      >
+        <Box display="flex">
+          <Text>{formatNationalId(user.nationalId)}</Text>
+        </Box>
+      </td>
+
+      <td
+        className={cn({
+          [`${styles.tablePadding} `]: true,
+        })}
+      >
+        <Text> {staffRoleDescription(user.roles)}</Text>
+      </td>
+
+      {isLoggedInUser === false && (
         <td
           className={cn({
             [`${styles.tablePadding} `]: true,
           })}
         >
-          <Box display="flex">
-            <Text>{formatNationalId(user.nationalId)}</Text>
-          </Box>
+          {user.active ? (
+            <Button variant="text" colorScheme="destructive">
+              Óvirkja
+            </Button>
+          ) : (
+            <Button variant="text" colorScheme="light">
+              Virkja
+            </Button>
+          )}
         </td>
-
-        <td
-          className={cn({
-            [`${styles.tablePadding} `]: true,
-          })}
-        >
-          <Text> {staffRoleDescription(user.roles)}</Text>
-        </td>
-
-        {isLoggedInUser === false && (
-          <td
-            className={cn({
-              [`${styles.tablePadding} `]: true,
-            })}
-          >
-            {user.active ? (
-              <Button variant="text" colorScheme="destructive">
-                Óvirkja
-              </Button>
-            ) : (
-              <Button variant="text" colorScheme="light">
-                Virkja
-              </Button>
-            )}
-          </td>
-        )}
-      </tr>
-    </>
+      )}
+    </tr>
   )
 }
