--- apps/financial-aid/web-veita/src/components/UsersTableBody/index.ts
@@ -1,1 +1,0 @@
-export { default as UsersTableBody } from './UsersTableBody'
