+++ apps/financial-aid/web-veita/src/components/index.ts
@@ -32,3 +32,5 @@ export { default as TimeLineContainer } from './Timeline/TimeLineContainer'
 export { default as UsersTableBody } from './UsersTableBody/UsersTableBody'
 export { default as ApplicationNotFound } from './ApplicationNotFound/ApplicationNotFound'
 export { default as ApplicationProfile } from './ApplicationProfile/ApplicationProfile'
+export { default as EmployeeSideNavItems } from './EmployeeSideNavItems/EmployeeSideNavItems'
+export { default as AdminSideNavItems } from './AdminSideNavItems/AdminSideNavItems'
